requirements_TPU.txt
# General packages
numpy==1.22.0
tqdm
diffusers==0.26.2
pillow==9.4.0

# Results
pandas==2.0.1
IPython==8.12.3
ipykernel==6.25.1
jupyter_client==8.3.1
jupyter_core==5.3.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
traitlets==5.9.0

